Prevent errors when pyperclip is not installed
If pyperclip is not installed, the user simply sees a message showing
that the module is not installed and copying does not work. The
generation and printing passphrases is unaffected.
This is done since someone does not need to install pyperclip in order
to run this software. They can run it without dependencies except when
they need to copy something